% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancel.test-package.R
\docType{package}
\name{cancel.test-package}
\alias{cancel.test}
\alias{cancel.test-package}
\title{cancel.test: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/krlmlr/duckdb-r}
  \item \url{https://krlmlr.github.io/duckdb-r}
  \item Report bugs at \url{https://github.com/krlmlr/duckdb-r/issues}
}

}
\author{
\strong{Maintainer}: Kirill Müller \email{kirill@cynkra.com} (\href{https://orcid.org/0000-0002-1416-3412}{ORCID})

}
\keyword{internal}
